+++ test/version_change_test.dart
@@ -4,9 +4,21 @@ import 'package:version_manipulation/mutations.dart';
 import 'package:version_manipulation/version_extension.dart';
 
 void main() {
-  final version = Version.parse('0.1.3+foo.1');
-  test('Set build', () {
-    expect(version.change(build: ['moo']).toString(), '0.1.3+moo');
+  final version = Version.parse('0.1.3-alpha+foo.1');
+  test('Change', () {
+    expect(version.change(build: ['moo']).toString(), '0.1.3-alpha+moo');
+    expect(version.change(preRelease: ['beta']).toString(), '0.1.3-beta+foo.1');
+    expect(version.change(major: 42).toString(), '42.1.3-alpha+foo.1');
+    expect(version.change(minor: 42).toString(), '0.42.3-alpha+foo.1');
+    expect(version.change(patch: 42).toString(), '0.1.42-alpha+foo.1');
+    expect(
+        version.change(
+            major: 42,
+            minor: 42,
+            patch: 42,
+            preRelease: ['gamma', 'omega'],
+            build: ['foo', 'bar', '42']).toString(),
+        '42.42.42-gamma.omega+foo.bar.42');
   });
   group('Build', () {
     test('bumping empty build sets it to 1', () {
@@ -42,6 +54,28 @@ void main() {
     test('KeepBuild', () {
       expect(KeepBuild(BumpBreaking())(Version.parse('0.2.3+foo')).toString(),
           '0.3.0+foo');
+      expect(KeepBuild(BumpBreaking())(Version.parse('0.2.3')).toString(),
+          '0.3.0');
+    });
+    test('KeepPreRelease', () {
+      expect(
+          KeepPreRelease(BumpBreaking())(Version.parse('0.2.3-alpha-foo.1.2.3'))
+              .toString(),
+          '0.3.0-alpha-foo.1.2.3');
+      expect(KeepPreRelease(BumpBreaking())(Version.parse('0.2.3')).toString(),
+          '0.3.0');
+    });
+    test('KeepPreRelease and KeepBuild', () {
+      expect(
+          KeepBuild(KeepPreRelease(BumpBreaking()))(
+                  Version.parse('0.2.3-alpha-foo.1.2.3+42'))
+              .toString(),
+          '0.3.0-alpha-foo.1.2.3+42');
+      expect(
+          KeepPreRelease(KeepBuild(BumpBreaking()))(
+                  Version.parse('0.2.3-alpha-foo.1.2.3+42'))
+              .toString(),
+          '0.3.0-alpha-foo.1.2.3+42');
     });
   });
 }