+++ CHANGELOG.md
@@ -4,8 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
-## 0.0.2 - 2020-07-18
+## [Unreleased]
+## [0.0.3] - 2020-09-01
+### Added
+- KeepPreRelease wrapper
+- `preRelease` argument to `change()`
+
+### Fixed
+- KeepBuild should not clear pre-release part
+
+## [0.0.2] - 2020-07-18
 ### Added
 - Mutations
 
@@ -13,3 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `nextBuild`
 - `set`
+
+[Unreleased]: https://github.com/f3ath/dart-version-manipulation/compare/0.0.3...HEAD
+[0.0.3]: https://github.com/f3ath/dart-version-manipulation/compare/0.0.2...0.0.3
+[0.0.2]: https://github.com/f3ath/dart-version-manipulation/compare/0.0.1...0.0.2